Remove repeat code
