+++ fusion/helper_scripts/sim_helpers.py
@@ -11,6 +11,35 @@ import numpy as np
 import yaml
 
 
+def log_message(message, log_queue):
+    """
+    Logs a message.
+    """
+    if log_queue:
+        log_queue.put(message)
+    else:
+        print(message)
+
+def update_matrices(info_dict: dict):
+    """
+    Misc. update of network resource matrices.
+    """
+    resp = {
+        'times_matrix': [],
+        'sims_matrix': [],
+        'networks_matrix': [],
+        'dates_matrix': [],
+    }
+    for curr_time, obj in info_dict.items():
+        resp['times_matrix'].append([curr_time])
+        resp['sims_matrix'].append(obj['sim_list'])
+        resp['networks_matrix'].append(obj['network_list'])
+        resp['dates_matrix'].append(obj['dates_list'])
+        resp['algorithms_matrix'].append(obj['algorithm_list'])
+
+    return resp
+
+
 def get_path_mod(mods_dict: dict, path_len: int):
     """
     Choose a modulation format that will allocate a network request.
